refactor : 유저 인증 모듈에 로그인 관련 클래스 이동 및 의존성 추가
- jwt 의존성 추가 ('io.jsonwebtoken:jjwt:0.9.1')
- 트랜잭션을 위해 spring-data-jpa 의존성 추가

의문점 - 인증 관련 interceptor 와 argumentResolver를 해당 모듈에 포함시켰습니다. 괜찮을지..?
그렇게 한 이유 - 해당 interceptor와 argumentResolver도 jwt와 결합되어 있기 때문에.
