+++ authorization-jwt/build.gradle
@@ -1,4 +1,7 @@
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'io.jsonwebtoken:jjwt:0.9.1'
+
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
