+++ authorization-jwt/src/main/java/com/momsitter/authiorization/exception/InvalidAuthTokenException.java
@@ -1,4 +1,6 @@
-package com.momsitter.exception;
+package com.momsitter.authiorization.exception;
+
+import com.momsitter.common.exception.UnauthorizedException;
 
 public class InvalidAuthTokenException extends UnauthorizedException {
     public InvalidAuthTokenException(String message) {
