+++ authorization-jwt/src/main/java/com/momsitter/authiorization/infrastructure/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package com.momsitter.infrastructure;
+package com.momsitter.authiorization.infrastructure;
 
 import io.jsonwebtoken.*;
 import org.springframework.beans.factory.annotation.Value;
