+++ authorization-jwt/src/main/java/com/momsitter/authiorization/ui/dto/auth/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.momsitter.ui.dto.auth;
+package com.momsitter.authiorization.ui.dto.auth;
 
 public class TokenResponse {
