+++ authorization-jwt/src/test/java/com/momsitter/Application.java
@@ -1,13 +1,13 @@
-package com.momsitter.authiorization;
+package com.momsitter;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class AuthiorizationJwtApplication {
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(AuthiorizationJwtApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 
 }
