--- authorization-jwt/src/test/java/com/momsitter/authiorization/AuthiorizationJwtApplicationTests.java
@@ -1,13 +1,0 @@
-package com.momsitter.authiorization;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuthiorizationJwtApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
