+++ authorization-jwt/src/test/resources/application.yml
@@ -1,0 +1,5 @@
+spring:
+  profiles:
+    active: test
+    include:
+      - auth
